scout: fix watch docs (#23506)
<!--Delete sections as needed -->

## Description

Fixes the generated scout watch docs to use examples from the upstream
markdown, instead of the yaml. When previously updated, only the
markdown was updated upstream and not the CLI source which generates the
yaml.

- Added new layout for the scout cli, based on the generic cli layout.
The new layout uses the more extensive markdown examples if present,
otherwise use the examples from the yaml.
- Updated the scout watch topic to use the new layout 

Once fixed upstream, will have a follow up to most likely revert this.



https://deploy-preview-23506--docsdocker.netlify.app/reference/cli/docker/scout/watch/#examples

## Related issues or tickets

Original PR to update the version: #23485. The markdown was updated but
not the yaml.
Upstream fix: https://github.com/docker/scout-cli-plugin/pull/713

## Reviews

<!-- Notes for reviewers here -->
<!-- List applicable reviews (optionally @tag reviewers) -->

- [ ] Editorial review

Signed-off-by: Craig Osterhout <craig.osterhout@docker.com>