engine/install: refresh debian.md for compliance with Deb822
Deprecate sources.list and use .sources instead

For context Deb822 format was introduced in apt version 1.1 ten years ago
and with the recent releases there is now a modernize-sources subcommand
to encourage migration into .sources format.

sources.list(5) says of the Deb822 ".sources" format "It is intended to
make this format gradually the default format, deprecating the previously
described one-line-style format, as it is easier to create, extend and
modify for humans and machines alike..."

https://wiki.debian.org/DebianRepository/UseThirdParty

Note for clarity /etc/apt/keyrings as the Docker install guide suggest is
in fact the correct location for a keyring installed by the system
operator, so the guide gets it half-right already.

Also some minor cleanup of the installation workflow where appropriate,
prefer sequoia over gpg, and prefer curl over wget, as appears in the
above Wiki article. Contextual examples are updated with the latest stable
release strings. Package name lists are sorted with `LANG=C sort`.